Move arrayOrSingleItemtoArray into src/core/types/array
