+++ src/types.js
@@ -36,19 +36,6 @@ export function isArray(value) {
 }
 
 /**
- * If the specified argument is an array, it's returned as is. If it's a
- * single item, the array containing this item is created and returned.
- * @param {!Array<T>|T} arrayOrSingleItem
- * @return {!Array<T>}
- * @template T
- */
-export function arrayOrSingleItemToArray(arrayOrSingleItem) {
-  return isArray(arrayOrSingleItem)
-    ? /** @type {!Array<T>} */ (arrayOrSingleItem)
-    : [arrayOrSingleItem];
-}
-
-/**
  * Determines if value is actually an Object.
  * @param {*} value
  * @return {boolean}
